Implemented the voyager library and a first navigation to the detail screen.
