+++ composeApp/src/commonMain/kotlin/detail/presentation/DetailScreen.kt
@@ -1,0 +1,59 @@
+package detail.presentation
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material.Icon
+import androidx.compose.material.IconButton
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.ArrowBack
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+import cafe.adriel.voyager.core.screen.Screen
+import cafe.adriel.voyager.navigator.LocalNavigator
+
+class DetailScreen : Screen {
+
+    @Composable
+    override fun Content() {
+
+        val navigator = LocalNavigator.current
+
+        Box(modifier = Modifier.fillMaxSize()) {
+
+            IconButton(
+                onClick = {
+                    navigator?.pop()
+
+                }, modifier = Modifier
+                    .padding(start = 14.dp, top = 14.dp)
+                    .clip(CircleShape)
+                    .background(MaterialTheme.colors.primary)
+            ) {
+                Icon(
+                    imageVector = Icons.Default.ArrowBack,
+                    contentDescription = null,
+                    tint = Color.White
+                )
+            }
+
+            Column(
+                verticalArrangement = Arrangement.Center,
+                horizontalAlignment = Alignment.CenterHorizontally,
+                modifier = Modifier.fillMaxSize(),
+            ) {
+                Text("Detail Screen")
+            }
+        }
+    }
+}
